requirements.txt
pygame==2.1.2
python-dotenv==0.20.0
